Updating HelloGDX to use the ApplicationFragment vs. activity
This makes it easier to support API level  28 which is deprecating
Fragment in favor of compat.Fragment.

Also added fragment at the beginning to handle the determination of
being ARCore compatible.  Then based on that determination, prompt the
user to use AR mode or 3D mode.

Based on that decision, another fragment is loaded either using ARCore
or not.

Also fixed the mis-rotated texture for Andy and added a check for
a zero point plane which I think was causing a buffer overflow.